Minor: rename `row_groups.rs` to `row_group_filter.rs`
